fix: 지원 시 파일을 첨부 하지 않았을 경우의 처리 추가

+++ application-service/src/main/java/club/gach_dong/service/ApplicationService.java
@@ -263,11 +263,13 @@ public class ApplicationService {
 
         List<ApplicationDocs> applicationDocsList = applicationDocsRepository.findAllByApplicationId(applyId);
 
-        for(ApplicationDocs applicationDocs : applicationDocsList){
-            objectStorageService.deleteObject(applicationDocs.getFileUrl());
-        }
+        if(!applicationDocsList.isEmpty()){
+            for(ApplicationDocs applicationDocs : applicationDocsList){
+                objectStorageService.deleteObject(applicationDocs.getFileUrl());
+            }
 
-        applicationDocsRepository.deleteAllByApplicationId(application.getId());
+            applicationDocsRepository.deleteAllByApplicationId(application.getId());
+        }
 
         applicationRepository.delete(application);
     }
